+++ contrib/packaging/startos/startos/utils.ts
@@ -82,6 +82,5 @@ export type GetBlockchainInfo = {
   chain: string
   blocks: number
   headers: number
-  verificationprogress: number
   initialblockdownload: boolean
 }